wasi_abi_fd.c: make integer truncations explicit

+++ libwasi/wasi_abi_fd.c
@@ -208,7 +208,7 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        uint32_t r = host_to_le32(n);
+        uint32_t r = host_to_le32((uint32_t)n);
         host_ret = wasi_copyout(ctx, wasi_memory(wasi), &r, retp, sizeof(r),
                                 WASI_U32_ALIGN);
         ret = 0;
@@ -267,7 +267,7 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        uint32_t r = host_to_le32(n);
+        uint32_t r = host_to_le32((uint32_t)n);
         host_ret = wasi_copyout(ctx, wasi_memory(wasi), &r, retp, sizeof(r),
                                 WASI_U32_ALIGN);
         ret = 0;
@@ -342,7 +342,7 @@ tty_hack:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        uint32_t r = host_to_le32(n);
+        uint32_t r = host_to_le32((uint32_t)n);
         host_ret = wasi_copyout(ctx, wasi_memory(wasi), &r, retp, sizeof(r),
                                 WASI_U32_ALIGN);
         ret = 0;
@@ -401,7 +401,7 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        uint32_t r = host_to_le32(n);
+        uint32_t r = host_to_le32((uint32_t)n);
         host_ret = wasi_copyout(ctx, wasi_memory(wasi), &r, retp, sizeof(r),
                                 WASI_U32_ALIGN);
         ret = 0;
